1.19.4

+++ src/main/java/net/mrscauthd/beyond_earth/common/entities/AlienZombieEntity.java
@@ -75,7 +75,7 @@ public class AlienZombieEntity extends Monster implements RangedAttackMob {
 
 	@Override
 	public boolean checkSpawnRules(LevelAccessor p_21686_, MobSpawnType p_21687_) {
-		BlockState blockState = level.getBlockState(new BlockPos(this.getX(), this.getY() - 1, this.getZ()));
+		BlockState blockState = level.getBlockState(new BlockPos((int)this.getX(), (int)this.getY() - 1, (int)this.getZ()));
 
 		if (blockState.is(Blocks.LAVA) || blockState.is(Blocks.AIR)) {
 			return false;
